requirements.txt
datasets==2.12.0
huggingface-hub==0.16.4
numpy==1.23.5
pandas==1.5.2
psutil==5.9.4
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tokenizers==0.13.3
#--find-links https://download.pytorch.org/whl/cu116
#torch==1.13.1+cu116 # use pip install torch==1.13.1+cu116 torchvision==0.14.1+cu116 torchaudio==0.13.1 --extra-index-url https://download.pytorch.org/whl/cu116
torch-tb-profiler==0.4.1
torchtext==0.14.1
torchvision==0.14.1
tqdm
transformers==4.27.4
nltk==3.6.7
wandb==0.14.2
scipy==1.10.1
spacy==3.7.2
scikit-learn==1.3.2